@@ -1,14 +1,15 @@
-import type { NextPage } from 'next';
+"use client"
+
 import React, { useRef, useEffect, useState } from 'react';
-import { useEntryStore } from '../hooks/useEntryStore';
-import { injectCode, reference_image, showPreview } from '../config/event';
+import { useRouter } from 'next/navigation';
+import { useEntryStore } from '../../hooks/useEntryStore';
+import { injectCode, reference_image, showPreview } from '../../config/event';
 
-import styles from '../styles/preview.module.scss';
-import { Modal } from '../components/modal';
-import { Button } from '../components/button';
-import { useRouter } from 'next/router';
+import {Modal} from "../../components/modal/Modal";
+import {Button} from "../../components/button/Button";
+import styles from '../../styles/preview.module.scss';
 
-const Home: NextPage = () => {
+export default function Page() {
   const { entry, updateIsSubmitted } = useEntryStore();
   const router = useRouter();
   const iframeRef = useRef<HTMLIFrameElement>(null);
@@ -55,6 +56,4 @@ const Home: NextPage = () => {
       )}
     </>
   );
-};
-
-export default Home;
+}
